+++ qa/testing.go
@@ -210,21 +210,6 @@ type server struct {
 	URL   string
 }
 
-type testCredentialsProvider struct {
-	token string
-}
-
-func (testCredentialsProvider) Name() string {
-	return "test"
-}
-
-func (t testCredentialsProvider) Configure(ctx context.Context, cfg *config.Config) (func(*http.Request) error, error) {
-	return func(r *http.Request) error {
-		r.Header.Set("Authorization", fmt.Sprintf("Bearer %s", t.token))
-		return nil
-	}, nil
-}
-
 func (f ResourceFixture) setupClient(t *testing.T) (*common.DatabricksClient, server, error) {
 	token := "..."
 	if f.Token != "" {
